refactor(polytope_playground): extract projection modes into projections module
Moves WireframeProjection + impl, the Schlegel parameter machinery, edge-blend selection, mode annotations, and hyperslice_cull_active (~385 lines) out of state.rs into projections.rs. state.rs re-exports them (a within-crate barrel) so impl Demo, the test module, and consumers are unchanged. clippy -D warnings clean, 97 tests pass.
